Fixed ComicInfo data-loss issue. Added analyzer.
The ComicInfo deserialization had an incorrect implementation, therefore
causing repaired chapters to lose information upon the re-save of the
ComicInfo file. This has been remedied, and an analyzer application was
added to detect damaged/missing meta-information in existing archives.
Replace the damaged archives with new ones to ensure your library is
perfect, again.